push api endpoint
This was done as part of a live pairing exercise months ago but I'm just
now checking in that work in progress.

The idea was to have an endpoint that a lighter-weight on-board client
could use for receiving commands. This would work better on something
like an Arduino where simple HTTP requests may be supported.